Fix compiler warnings across Netdata C code (#23998)
* fix: clean up compiler warnings in netdata code

glibc 2.43 makes strchr/strrchr/strstr return const char * for a const
argument, which turned several read-only pointers into
-Wdiscarded-qualifiers warnings. Make those pointers const, and give
is_filename() a non-const argument since it writes into its buffer.

Also: pass a non-const callback array to mnl_cb_run2() (older libmnl
lacks the const), drop the waitq NO_PRIORITY alias that clashed with
winspool.h, spell 'RSMB' without a multi-character constant, cast frame
pointers through uintptr_t, and rename the dbengine local named main.

No behaviour change.

* fix(aral): guard stats in the used/padding helpers like their siblings

aral_structures_bytes_from_stats() and aral_free_bytes_from_stats()
return 0 for a NULL stats pointer, but the used and padding helpers did
not. Inlined together in registry_statistics(), gcc 16 LTO took the
sibling checks as a possible NULL and flagged the unguarded loads with
-Wstringop-overflow. stats is never NULL, so this only makes the four
helpers consistent; no behaviour change.

* fix: remove dead counters and signedness mismatches flagged by gcc 16

gcc 16 counts a bare x++ as not a use, so a dev-mode build (-Wall -Wextra
-Wformat-signedness) flags many write-only locals: debug counters and
accumulators that nothing reads. Remove them. Keep two with
__maybe_unused: progress.c's archived (read by an assert) and
proc_stat.c's cpu_wakeups.

Also match four format specifiers to their argument types; the printed
text is unchanged, including the streaming CLABEL line.

No behaviour change.

* fix: keep -1 for unselected weights indexes; mark unused health_log arg

In the weights JSON the context/instance index is ssize_t with -1 for
"not selected", but `x ? dun->i : -1` has uint32_t type, so the -1 became
4294967295 before assignment. Cast the index arm to ssize_t.

health_create_alarm_entry() never uses its host argument; mark it
__maybe_unused rather than touch its five callers.

+++ src/libnetdata/stacktrace/stacktrace-libbacktrace.c
@@ -287,7 +287,7 @@ void impl_stacktrace_to_buffer(STACKTRACE trace, BUFFER *wb) {
             buffer_putc(wb, '#');
             buffer_print_uint64(wb, i);
             buffer_strcat(wb, " <unknown> [");
-            buffer_print_uint64_hex(wb, (uint64_t)st->frames[i]);
+            buffer_print_uint64_hex(wb, (uint64_t)(uintptr_t)st->frames[i]);
             buffer_putc(wb, ']');
         }
     }
